+++ packages/core/src/cleaners.ts
@@ -1,0 +1,20 @@
+export function deleteUndefinedValues<T extends Record<string, any>>(o: T): T {
+    if (typeof o === "object")
+        for (const k in o) if (o[k] === undefined) delete o[k]
+    return o
+}
+
+export function deleteEmptyValues<T extends Record<string, any>>(o: T): T {
+    if (typeof o === "object")
+        for (const k in o) {
+            const v = o[k]
+            if (
+                v === undefined ||
+                v === null ||
+                v === "" ||
+                (Array.isArray(v) && !v.length)
+            )
+                delete o[k]
+        }
+    return o
+}
